requirements.txt
pandas
streamlit
openai
streamlit-authenticator
pyyaml
yamlloader
